dev-requirements.txt
pytest
pytest-cov
ckanapi
black
flake8